+++ internal/provider/resource_iosxe_interface_ethernet_test.go
@@ -48,6 +48,8 @@ func TestAccIosxeInterfaceEthernet(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.address", "10.10.10.10"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.global", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.vrf", "VRF1"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "source_template.0.template_name", "TEMP1"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "source_template.0.merge", "false"))
 	if os.Getenv("IOSXE179") != "" || os.Getenv("IOSXE1712") != "" {
 		checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "bfd_template", "bfd_template1"))
 	}
@@ -82,8 +84,8 @@ func TestAccIosxeInterfaceEthernet(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_flow_monitors.0.name", "MON1"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "ip_flow_monitors.0.direction", "input"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "load_interval", "30"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "snmp_trap_link_status", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "logging_event_link_status_enable", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "snmp_trap_link_status", "false"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "logging_event_link_status_enable", "false"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -142,13 +144,20 @@ resource "iosxe_restconf" "PreReq3" {
 	depends_on = [iosxe_restconf.PreReq2, ]
 }
 
+resource "iosxe_restconf" "PreReq4" {
+	path = "Cisco-IOS-XE-native:native/template/Cisco-IOS-XE-template:template_details=TEMP1"
+	attributes = {
+		"template_name" = "TEMP1"
+	}
+}
+
 `
 
 func testAccIosxeInterfaceEthernetConfig_minimum() string {
 	config := `resource "iosxe_interface_ethernet" "test" {` + "\n"
 	config += `	type = "GigabitEthernet"` + "\n"
 	config += `	name = "3"` + "\n"
-	config += `	depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, iosxe_restconf.PreReq2, iosxe_restconf.PreReq3, ]` + "\n"
+	config += `	depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, iosxe_restconf.PreReq2, iosxe_restconf.PreReq3, iosxe_restconf.PreReq4, ]` + "\n"
 	config += `}` + "\n"
 	return config
 }
@@ -174,6 +183,10 @@ func testAccIosxeInterfaceEthernetConfig_all() string {
 	config += `		address = "10.10.10.10"` + "\n"
 	config += `		global = false` + "\n"
 	config += `		vrf = "VRF1"` + "\n"
+	config += `	}]` + "\n"
+	config += `	source_template = [{` + "\n"
+	config += `		template_name = "TEMP1"` + "\n"
+	config += `		merge = false` + "\n"
 	config += `	}]` + "\n"
 	if os.Getenv("IOSXE179") != "" || os.Getenv("IOSXE1712") != "" {
 		config += `	bfd_template = "bfd_template1"` + "\n"
@@ -215,9 +228,9 @@ func testAccIosxeInterfaceEthernetConfig_all() string {
 	config += `		direction = "input"` + "\n"
 	config += `	}]` + "\n"
 	config += `	load_interval = 30` + "\n"
-	config += `	snmp_trap_link_status = true` + "\n"
-	config += `	logging_event_link_status_enable = true` + "\n"
-	config += `	depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, iosxe_restconf.PreReq2, iosxe_restconf.PreReq3, ]` + "\n"
+	config += `	snmp_trap_link_status = false` + "\n"
+	config += `	logging_event_link_status_enable = false` + "\n"
+	config += `	depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, iosxe_restconf.PreReq2, iosxe_restconf.PreReq3, iosxe_restconf.PreReq4, ]` + "\n"
 	config += `}` + "\n"
 	return config
 }